+++ functions/lib/db/rooms.js
@@ -33,6 +33,11 @@ class Rooms extends DbAccess {
     return `${appUrl}/join_room?otp=${invite.otp}&iid=${invite.id}&email=${email}`
   }
 
+  async getClaimUrl(appUrl, claim) {
+    const email = util.args.consolidateEmailString(claim.email)
+    return `${appUrl}/claim_room?otp=${claim.otp}&cid=${claim.id}&email=${email}`
+  }
+
   async inviteById(id) {
     return await db.pg.massive.room_invitations.findOne({id: id})
   }
@@ -68,7 +73,7 @@ class Rooms extends DbAccess {
       ORDER BY
         rooms.id ASC
     `)
-    return this.namedRoomsListToMostPreferredList(namedRooms)
+    return this.namedRoomsListToMostPreferredList(namedRooms)[0]
   }
 
   async isMember(userId, roomId) {
@@ -162,6 +167,53 @@ class Rooms extends DbAccess {
     })
   }
 
+  async tryToCreateClaim(email, roomName, allowRegistered=false, createNewRooms=false) {
+    email = lib.util.args.consolidateEmailString(email)
+    const user = await db.pg.massive.users.findOne({ email: email })
+    if(user) {
+      if(allowRegistered) {
+        console.log(`Warning: user ${user.email} already registered.`)
+      } else {
+        return { error: lib.db.ErrorCodes.user.ALREADY_REGISTERED, user: user }
+      }
+    }
+
+    let roomNameEntry = await db.pg.massive.room_names.findOne({name: roomName})
+    if(!roomNameEntry) {
+      if(createNewRooms) {
+        const priorityLevel = 1
+        const isVanity = true
+        const ownerId = null
+        const createResult =  await this.createRoomWithName(roomName, ownerId, priorityLevel, isVanity)
+        roomNameEntry = createResult.nameEntry
+      } else {
+        return { error: lib.db.ErrorCodes.room.UNKNOWN_ROOM }
+      }
+    }
+
+    let room = await this.roomById(roomNameEntry.room_id)
+    const existingClaim = await db.pg.massive.room_claims.findOne({room_id: room.id})
+    if(existingClaim) {
+      return { error: lib.db.ErrorCodes.room.CLAIM_UNIQUENESS }
+    }
+    const claim = await this.createClaim(room.id, email)
+    return { claim }
+  }
+
+  async createClaim(roomId, claimerEmail, expiresAt = null) {
+    return await db.pg.massive.room_claims.insert({
+      room_id: roomId,
+      email: claimerEmail,
+      otp: lib.db.otp.generate(),
+      issued_at: this.now(),
+      expires_at: expiresAt
+    })
+  }
+
+  async claimUpdateEmailedAt(claimId) {
+    return await db.pg.massive.room_claims.update({id: claimId}, {emailed_at: this.now()})
+  }
+
   isValidInvitation(invitation, email, otp) {
     const verification = lib.db.otp.verify(invitation, otp)
     if(verification.error != null) {
@@ -203,6 +255,24 @@ class Rooms extends DbAccess {
     }
   }
 
+  async resolveClaim(claim, user, otp) {
+    const verification = this.isValidInvitation(claim, user.email, otp)
+    if(verification.error != null) {
+      return verification
+    }
+
+    try {
+      return await db.pg.massive.withTransaction(async (tx) => {
+        await tx.room_claims.update({id: claim.id}, {resolved_at: this.now()})
+        return await tx.rooms.update({id: claim.room_id}, {owner_id: user.id})
+      })
+
+    } catch(e) {
+      // TODO: ERROR_LOGGING
+      return { error: lib.db.ErrorCodes.UNEXPECTER_ERROR }
+    }
+  }
+
 
   async tryToGenerateRoom(userId) {
     const canGenerate = await this.underMaxOwnedRoomLimit(userId)
@@ -223,14 +293,19 @@ class Rooms extends DbAccess {
       idString = this.generateIdString()
       isUnique = await this.isUniqueIdString(idString)
     }
+    return await this.createRoomWithName(idString, userId)
+  }
+
+  async createRoomWithName(name, ownerId=null, priorityLevel = 0, isVanity=false) {
     return await db.pg.massive.withTransaction(async (tx) => {
       const room = await tx.rooms.insert({
-        owner_id: userId
+        owner_id: ownerId
       })
       const generatedName = await tx.room_names.insert({
         room_id: room.id,
-        name: idString,
-        priority_level: 0
+        name: name,
+        priority_level: priorityLevel,
+        is_vanity: isVanity
       })
       return {
         room: room,